faster vops
Where a vector is initalized and then overwritten with other values, I previously used vec![0.; n]. This performs an allocation with zeros, but the zeros aren't actually used for anything. The new way uses with_capacity and (safe) set_len, which fills the vector with garbage (which doesn't matter because it's all getting overwritten anyways). This "allocation" is free.